requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
motor
python-dotenv
psutil
aiofiles
